Fix related to events inAPI

+++ common/types/activation.go
@@ -181,13 +181,13 @@ type ActivationTx struct {
 }
 
 // NewActivationTx returns a new activation transaction. The ATXID is calculated and cached.
-func NewActivationTx(challenge NIPostChallenge, coinbase Address, nipost *NIPost, numUnits uint, initialPost *Post) *ActivationTx {
+func NewActivationTx(challenge NIPostChallenge, coinbase Address, nipost *NIPost, numUnits uint64, initialPost *Post) *ActivationTx {
 	atx := &ActivationTx{
 		InnerActivationTx: InnerActivationTx{
 			ActivationTxHeader: ActivationTxHeader{
 				NIPostChallenge: challenge,
 				Coinbase:        coinbase,
-				NumUnits:        uint64(numUnits),
+				NumUnits:        numUnits,
 			},
 			NIPost:      nipost,
 			InitialPost: initialPost,
